Fixes and minor improvements, build update
General
* update build system:
  * replace Visual Studio project with CMake for Windows;
  * add simple packaging;
  * add option to use SDL_main.
* reformat LICENSE file with Markdown;
* update README.

Settings
* add settings update callback;
* add options menu;
* add automatic naming of log and replay files;
* add automatic removal of old log and replay files;
* remove 'release paths' for settings (wasn't even tested);
* set target_fps to display refresh rate by default.

Game
* add survival game mode;
* add player acceleration limit cancel if no enemies are nearby;
* make assembler objects in factories more distributed between lines;
* fix EC_Health::on_dmg not being invoked on killing shot;
* fix death message not appearing (duplicate check in PlayerManager::get_ent);
* update tutorial lines (add of mention laser designator).

Game internals
* change game mode handling to support more modes;
* GameControl - terrain lifetime is now bound to controller;
* rename LevelControl::ref_room(pos) to get_room;
* make all references to LevelCtrRoom const - mutable values are now marked as such;
* replace FixtureInfo with b2Fixture in CollisionEvent;
* make PhysicsWorld::query_aabb return last callback result;
* fix rifle grenade colliding with barrage sensor;
* increase rifle bullet speed;
* increase rocket projectile radius for player;
* move ERespawnFunc to common objects (used in survival mode);
* AI fixes:
  * add battle timeout hack (to reset turrets);
  * add AI_Drone::mut_pars() for updating parameters;
  * make pathfinding raycast check for all static objects.
* increase hacker placement distance from terminal;

Game UI
* LevelMap - add scrolling, marking visited maps, marking factories;
* add FadeText spread multiplier (for TutorialDummy);
* fix teleport menu (was broken in previous commit);

Utilities
* add thread naming function (for debugging);
* add LineCfg value bounds checks;
* fix serializer header writer for Clang;
* VIG bugfix:
  * warnboxes ignoring clicks and keypresses;
  * text formatting using vas_log (noticeable on sliders);
  * different vig_selector signature in source and header.
* fix rnd_stat() generator not being thread_local;
* replace most usages of TimeSpan::since_start() with TimeSpan::current();
* minor API and comment improvements in vas_file.

+++ scripts/load_deps.ps1
@@ -1,8 +1,9 @@
 # PowerShell downloader for MSVC
 
 $Git="git.exe"
-$TmpDir="load_deps_tmp"
-$TmpZip="load_deps_tmp.zip"
+$TmpDir="TMP_load_deps"
+$TmpZip="TMP_load_deps.zip"
+$ErrorActionPreference = 'Inquire'
 
 
 
@@ -16,62 +17,16 @@ if (!(Test-Path "core\main.cpp"))
 	Set-Location ".."
 }
 
-$GitVars="$Git","git.exe","C:\Program Files\Git\cmd\git.exe","C:\Program Files\Git\bin\git.exe"
-$Git=""
-
-ForEach ($Var in $GitVars) {
-	if ((Get-Command "$Var" -ErrorAction SilentlyContinue) -ne $null)
-	{
-		$Git="$Var"
-		break
-	}
+$OutDir="$(Get-Location)"
+if (Test-Path "$OutDir/external") {
+	Remove-Item -Recurse -Force "$OutDir/external"
 }
-
-if ("$Git") {
-	Write-Host "Found git at $Git"
-}
-else {
-	Write-Host "Unable to find git"
-	exit 1
-}
-
-$OutDir=Get-Location
-Remove-Item -Recurse -Force "$OutDir/external"
 New-Item -ItemType Directory -Path "$OutDir" -Name "external"
 
 $OutDir="$OutDir/external"
 New-Item -ItemType Directory -Force -Path "$OutDir" -Name "include"
 New-Item -ItemType Directory -Force -Path "$OutDir" -Name "msvc32"
 New-Item -ItemType Directory -Force -Path "$OutDir" -Name "msvc64"
-
-
-
-function dlsrc($name, $SrcEnd, $files)
-{
-	New-Item -ItemType Directory -Name "$TmpDir"
-	cd "$TmpDir"
-
-	&"$Git" init
-	&"$Git" config core.sparsecheckout true
-	&"$Git" remote add -f origin "$name"
-
-	$List=$files -join "`n" | Out-String
-	
-	"$List" | Out-File -Encoding ASCII -FilePath ".git/info/sparse-checkout"
-	&"$Git" pull origin master
-
-	if ($SrcEnd -eq 1) {
-		ForEach ($Var in $files) {
-			Move-Item -Path "$Var" -Destination "$OutDir"
-		}
-		dlsrc_end
-	}
-}
-function dlsrc_end()
-{
-	cd ..
-	Remove-Item -Recurse -Force "$TmpDir"
-}
 
 
 
@@ -89,7 +44,7 @@ function unzip($file, $dest)
 
 function dlzip($url, $include, $lib32, $lib64)
 {
-	Invoke-WebRequest -Uri "$url" -OutFile "$TmpZip" -UserAgent [Microsoft.PowerShell.Commands.PSUserAgent]::FireFox
+	Invoke-WebRequest -Uri "$url" -OutFile "$TmpZip"
 	
 	New-Item -ItemType Directory -Name "$TmpDir"
 	unzip "$TmpZip" "$TmpDir/"
@@ -110,19 +65,7 @@ function dlzip($url, $include, $lib32, $lib64)
 
 
 
-dlsrc "https://github.com/nothings/stb.git" 1 ("stb_image.h", "stb_image_write.h")
-
-dlsrc "https://github.com/erincatto/Box2D.git" 0 ("include", "src")
-Move-Item -Path "include/box2d" -Destination "$OutDir"
-Move-Item -Path "src/*" -Destination "$OutDir/box2d"
-dlsrc_end
-
-dlsrc "https://github.com/fmtlib/fmt.git" 0 ("include/fmt", "src/*.cc")
-Move-Item -Path "include/fmt" -Destination "$OutDir"
-Move-Item -Path "src/*.cc" -Destination "$OutDir/fmt"
-dlsrc_end
-
-dlzip "https://sourceforge.net/projects/glew/files/glew/2.1.0/glew-2.1.0-win32.zip/download" `
+dlzip "https://github.com/nigels-com/glew/releases/download/glew-2.1.0/glew-2.1.0-win32.zip" `
 	"glew*/include/GL" `
 	("glew*/lib/Release/Win32/glew32.lib", "glew*/bin/Release/Win32/glew32.dll") `
 	("glew*/lib/Release/x64/glew32.lib", "glew*/bin/Release/x64/glew32.dll")
@@ -132,10 +75,10 @@ dlzip "https://github.com/ubawurinna/freetype-windows-binaries/archive/master.zi
 	"freetype*/win32/*.*" `
 	"freetype*/win64/*.*"
 
-dlzip "https://www.libsdl.org/release/SDL2-devel-2.0.10-VC.zip" `
+dlzip "https://www.libsdl.org/release/SDL2-devel-2.0.12-VC.zip" `
 	"SDL2*/include" `
-	"SDL2*/lib/x86/SDL2.*" `
-	"SDL2*/lib/x64/SDL2.*" `
+	("SDL2*/lib/x86/SDL2.*", "SDL2*/lib/x86/SDL2main.*") `
+	("SDL2*/lib/x64/SDL2.*", "SDL2*/lib/x64/SDL2main.*")
 
-Move-Item -Path "$OutDir/include/include" -Destination "$OutDir/include/SDL2"
+Rename-Item "$OutDir/include/include" "$OutDir/include/SDL2"
